requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
yfinance>=0.2.36
feedparser>=6.0.10
beautifulsoup4>=4.12.0
requests>=2.31.0
nltk>=3.8.1
pandas>=2.1.0
apscheduler>=3.10.4
sqlalchemy>=2.0.25
python-dotenv>=1.0.0
google-genai>=1.0.0
